move run shield in separate file
